requirements.txt
flask
flask-cors
requests
google-generativeai 